Added refactor of subscriptions service, new method to mutate the file to not include the subscription ID and we always replace the subscrition Id. Did some changes to the file structure on advisor aswell.
