+++ src/steps/resource-manager/authorization/index.test.ts
@@ -39,10 +39,10 @@ import {
   RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
 import { filterGraphObjects } from '../../../../test/helpers/filterGraphObjects';
-import { fetchSubscription } from '../subscriptions';
 import { fetchKeyVaults } from '../key-vault';
 import { fetchManagementGroups } from '../management-groups';
 import { ManagementGroupEntities } from '../management-groups/constants';
+import { fetchSubscription } from '../subscriptions/executionHandlers/subscriptions';
 
 let recording: Recording;
